api: Update game search queries for twitch api v5 and v6
There is a change in TwitchAPIv5 for game search. Now it uses
query keyword without game name specification or language.

There is a change in TwitchAPIv6 for game search. Not it is
not guaranteed that search query returns 100 streams at max.
Application must fetching new data until query returns 0 streams.